New Features:
ModuleUpgradable added so when looking at an upgrade in the room it will display: cost per material, cost of an upgrade in total, and the estimated daily maintenance in denari.

ModuleIndustry has a 3rd method of displaying profit - the weighted average of revenue & saved values based on consumption / production of the resource. This should be the most accurate *current* profit per person, but it is likely that the next person will be getting "revenue" rates (sale).

Minor changes:

Node_extra edited where Production worker benefit shows up.